When a gen times out, shutdown the executor and wait for it
Because AP uses threadpools internally, when a gen times out, some of
these threads might still be going and continue printing stuff out which
would then be lost. By waiting for the executor to close up and
recreating it, we prevent this from happening. This is obviously not
ideal but a timeout case is not the common one so it should kill perfs
too much