Set some global undocumented settings
Main one here is `no_gui` which makes it so that if a ROM is missing it
won't open `worker_number` amounts of file pickers.
While I was there, also set `skip_autosave` since we don't want that,
I'm not expecting it to make a difference performance wise but it's
still better.