Fix deadlock if a hook had a faulty reclassify_outcome
If reclassify_outcome raised any exception, it would stop the timeout
handler without even classifying the first timeout, leading to the
fuzzer thinking that something was still generating even though the
generation was long gone, leading to a deadlock.

Found this with the ut hook because their `reclassify_outcome` assumed
that `after_generate` would always be called. It's a bug in the hook but
we should really not deadlock for that. It will now print the exception
in the logs when that happens and log the exception. I think there's a
chance for this to mess up if we traceback while dumping the generation
output or in `gen_callback` but those are not user controlled and would
be fuzzer bugs anyway so I'm fine with that.