Don't let timeout fire if a hook raises an exception in after_generate
Since 1a924916713ef2d4dec21da0a46359fd12d611b6, after_generate is
comprised in the timeout time but if a hook was to raise an exception in
here, we would skip cancelling the timer which could lead to it firing
*after* while we're in the error callback and that would make the fuzzer
report a failure, potentially delete the YAMLs, then try to report a
timeout, and fail to do so because the YAMLs are gone, then try again
because it would assume that a hook raised in `reclassify_outcome`.
Bonus point if said hook marks the result as ignored for a triple
classification of a single failure, none of which is correct.