Add hook for running an item collection accessibility test
Archipelago requires that collecting an item must only ever increase
accessibility or have no effect on accessibility. If accessibility
reduces when an item is collected, it can cause generation to crash or
behave in a number of weird ways.

Barring core bugs, failing the test in this hook means there is 100% a
logic issue in the world.